+++ lib/entity.ts
@@ -3,8 +3,17 @@
 
 import { EventEmitter } from "events";
 
+/**
+ * Abstract base class for all the entities like Connection, Session, Sender, Receiver in the
+ * AMQP protocol.
+ * @class Entity
+ */
 export abstract class Entity extends EventEmitter {
-  isBeingCreated: boolean = false;
+  /**
+   * @property actionInitiated Indicates whether the action of creating or closing an entity has
+   * been initiated. Default value: `false`.
+   */
+  actionInitiated: boolean = false;
   constructor() {
     super();
   }