+++ lib/link.ts
@@ -19,7 +19,7 @@ export enum LinkType {
 export abstract class Link extends Entity {
   linkOptions?: LinkOptions;
   type: LinkType;
-  isBeingCreated: boolean = false;
+  actionInitiated: boolean = false;
   protected _link: link;
   protected _session: Session;
   constructor(type: LinkType, session: Session, link: link, options?: LinkOptions) {
@@ -229,6 +229,7 @@ export abstract class Link extends Entity {
 
         const removeListeners = () => {
           clearTimeout(waitTimer);
+          this.actionInitiated = false;
           this._link.removeListener(errorEvent, onError);
           this._link.removeListener(closeEvent, onClose);
         };
@@ -260,6 +261,7 @@ export abstract class Link extends Entity {
         waitTimer = setTimeout(actionAfterTimeout,
           this.connection.options!.operationTimeoutInSeconds! * 1000);
         this._link.close();
+        this.actionInitiated = true;
       } else {
         resolve();
       }